+++ flowdyn/solution/euler_nozzle.py
@@ -11,53 +11,98 @@ import aerokit.aero.Isentropic as Is
 import aerokit.aero.ShockWave  as sw
 import aerokit.aero.MassFlow   as mf
 import aerokit.aero.nozzle     as nz
+import aerokit.common.defaultgas as defg
 
 class nozzle():
+    """ Define a nozzle
 
-    def __init__(self, model, section):
+    :param model: define a `modelphy.euler` model, needs the gamma coefficient
+    :param section: array of section law
+    :param NPR: NPR value (>1)
+    :param ref_rttot: additional definition of r*Ttot to complete state (default 1.)
+    :param scale_ps: arbitrary scaling of static (and associated total) pressure (default 1. at the outlet)
+    """
+
+    def __init__(self, model, section, NPR=None, ref_rttot=1., scale_ps=1.):
 
         self.model   = model
+        self._gam    = self.model.gamma
         self.section = section
-        self.AsoAc   = section[-1]/np.min(section)
+        self.ithroat = section.argmin()
+        self.AsoAc   = section[-1]/section[self.ithroat]
         self.NPR0, self.NPRsw, self.NPR1, self.Msub, self.Msh, self.Msup = nz._NPR_Ms_list(self.AsoAc)
-
-    def set_NPR(NPR):
-        if NPR < NPR0:
-            _Ms = Is.Mach_PiPs(NPR, gamma=self.model.gamma)
-            self._M  = mf.MachSub_Sigma(self.AsoAc*mf.Sigma_Mach(Ma_col)/self.AsoAc[-1], gamma=self.model.gamma)
-            self._Pt = 0.*coord_x + 1.
-            self._Ps = _Pt/Is.PiPs_Mach(self._M, gamma=self.model.gamma)
-        elif NPR < NPRsw:
-            _M  = mf.Mach_Sigma(Noz_AoAc, Mach=_Minit)
-            #
-            # analytical solution for Ms, losses and upstream Mach number of shock wave
-            Ms     = nz.Ms_from_AsAc_NPR(target_AoAc, NPR)
-            Ptloss = Is.PiPs_Mach(Ms)/NPR
-            Msh    = sw.Mn_Pi_ratio(Ptloss)
-            #
-            # redefine curves starting from 'ish' index (closest value of Msh in supersonic flow)
-            ish    = np.abs(_M-Msh).argmin()
-            _M[ish:] = mf.MachSub_Sigma(Noz_AoAc[ish:]*mf.Sigma_Mach(Ms)/target_AoAc)
-            _Pt[ish:] = Ptloss
-            _Ps = _Pt/Is.PiPs_Mach(_M)
-        #
-
-    def Mach(self):
+        self._ref_rttot = ref_rttot
+        self._scale_ps  = scale_ps
+        if NPR:
+            self.set_NPR(NPR)
         return
 
-    def primdata(self, ):
-       return [q.rho, q.u, q.p]
-        
-    def consdata(self, ):
-        return self.model.prim2cons([q.rho, q.u, q.p])
+    def set_NPR(self, NPR):
+        """ Define Nozzle Pressure Ratio (inlet Ptot over outlet Ps) for this case
 
-    def fdata(self, mesh, t=None):
-        qcons = self.model.prim2cons([q.rho, q.u, q.p])
+        :param NPR: NPR value (>1)
+
+        """
+        self.NPR = NPR
+        defg.set_gamma(self._gam)
+        if NPR < self.NPR0: # flow is fully subsonic
+            _Ms = Is.Mach_PiPs(NPR)
+            _M  = mf.MachSub_Sigma(self.section*mf.Sigma_Mach(_Ms)/self.section[-1])
+            _Pt = 0.*_M + NPR
+            _Ps = _Pt/Is.PiPs_Mach(_M)
+        elif NPR < self.NPRsw: # throat is choked, there may be a shock
+            # compute Mach, assumed to be subsonic before throat, supersonic after
+            _Minit = 0.*self.section +.5
+            _Minit[self.ithroat:] = 2.
+            _M  = mf.Mach_Sigma(self.section/self.section[self.ithroat], Mach=_Minit)
+            _Pt = NPR+0.*_M
+            # CHECK, there is a shock
+            # analytical solution for Ms, losses and upstream Mach number of shock wave
+            Ms     = nz.Ms_from_AsAc_NPR(self.AsoAc, NPR)
+            Ptloss = Is.PiPs_Mach(Ms)/NPR
+            Msh    = sw.Mn_Pi_ratio(Ptloss)
+            # redefine curves starting from 'ish' index (closest value of Msh in supersonic flow)
+            ish       = np.abs(_M-Msh).argmin()
+            _M[ish:]  = mf.MachSub_Sigma(self.section[ish:]*mf.Sigma_Mach(Ms)/self.section[-1])
+            _Pt[ish:] = Ptloss*NPR
+            _Ps = _Pt/Is.PiPs_Mach(_M)
+        #
+        self._M  = _M
+        self._Pt = _Pt
+        self._Ps = _Ps
+        return
+
+    def Mach(self):
+        """ """
+        return self._M
+
+    def Ptot(self):
+        """ """
+        return self._Ptot
+
+    def Ps(self):
+        """ """
+        return self._Ps
+
+    def primdata(self):
+        """ computes list of rho, u, p data
+
+        """
+        p   = self._scale_ps * self._Ps
+        rt  = self._ref_rttot/Is.TiTs_Mach(self._M, gamma=self._gam)
+        rho = p/rt
+        u   = self._M*np.sqrt(self._gam*rt) 
+        return [rho, u, p]
+        
+    def consdata(self):
+        """
+        """
+        return self.model.prim2cons(self.primdata())
+
+    def fdata(self, mesh):
+        """
+        """
+        qcons = self.consdata()
         return field.fdata(self.model, mesh, qcons)
 
-    def bcL(self):
-        return [self.rhoL, self.uL, self.pL]
-        
-    def bcR(self):
-        return [self.rhoR, self.uR, self.pR]