+++ CHANGELOG.md
@@ -16,11 +16,23 @@ pip show flowdyn
 
 ## Versions
 
-### [1.0.1](https://pypi.org/project/flowdyn/) (2021-01-27)
+### [1.x.x](https://pypi.org/project/flowdyn/) (2021-xx-xx)
 
-#### Bug fix
+#### changed
+
+- avoid warnings with `vanalbada` and `vanleer` limiters when uniform flows
+- analytical 1D solution for nozzle flows in `solution.euler_nozzle`
+- improve test coverage
+- optimize some mesh computation
+
+#### fixed
+
+### [1.0.1](https://pypi.org/project/flowdyn/) (2021-01-29)
+
+#### fixed
 
 - fix computation of `modelphy.euler` supersonic inlet condition 
+
 ### [1.0.0](https://pypi.org/project/flowdyn/) (2021-01-27)
 
 - models: convection, Burgers, Euler and derived (nozzle)